backend/pyproject.toml
[project]
name = "interview-python-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.2",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "langchain>=0.3.0",
    "langchain-ollama>=0.2.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.9",
]
